requirements.txt
flask
flask-pymongo
python-slugify
python-dotenv
python-dateutil
requests
apscheduler
